refactor: extract JiraExportProvider+Duplicates (#638 slice C)
Third slice of #638 (after #874 Networking, #890 ADF). Sibling of
GitHub's #881.

Moved (7 methods + 4 wire types):
- func findOpenIssues — already public, unchanged
- makeSearchRequest — stays private (single-caller)
- makeExportFingerprintSearchRequest — stays private (single-caller)
- existingExportResult — widen private → internal (called from export in base)
- reconcilePendingExport — widen private → internal (called from createWithRetry in base + existingExportResult moves with)
- findExportedIssue — stays private (single-caller)
- searchJQL — stays private (single-caller)
- JiraSearchRequest, JiraSearchResponse, JiraSearchIssue, JiraSearchIssueFields — file-scope private → internal

Stored-property widenings (found by codex + agy pre-review — Jira
actor's session/receiptStore were still private, unlike GitHub which
was widened by #880):
- session: private let → let (used by findOpenIssues + findExportedIssue)
- receiptStore: private let → let (used by existingExportResult + reconcilePendingExport)

Line-count delta:
- JiraExportProvider.swift: 1033 → 850 (-183 lines)
- JiraExportProvider+Duplicates.swift (new): 197 lines
- xcodeproj: +4

Test coverage: JiraExportProviderTests passes unchanged.

Closes #892. Continues #638 after PRs #877/#891.
